app/requirements.txt
boto3==1.34.162
pydantic==2.9.1
pydantic-settings==2.5.2
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==23.0.0
psycopg2-binary==2.9.9
pydantic==2.9.1
bcrypt==4.1.2
PyJWT==2.9.0
boto3==1.34.162
python-dotenv==1.0.1
redis==5.0.7
email-validator==2.2.0
httpx==0.27.2
